retira contabilização de impressões do model e leva para action (helper); retira acentuação de chave de array
