+++ app/Helpers/PrintingHelper.php
@@ -2,6 +2,9 @@
 
 namespace App\Helpers;
 
+use App\Models\User;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Symfony\Component\Process\Process;
 use Symfony\Component\Process\Exception\ProcessFailedException;
@@ -168,5 +171,45 @@ class PrintingHelper
 
         if (!$process->isSuccessful()) return '';
         return $pdf;
+    }
+
+    /**
+     * Função para retornar a quantidade de impressões em determinado contexto.
+     *
+     * @param string  $user    N.USP
+     * @param Printer $printer objeto Printer
+     * @param string  $period  Mensal ou diário
+     *
+     * @return int quantidade de impressões para o contexto
+     */
+    public static function getPrintingsQuantities(User $user = null, $printer = null, $period = null)
+    {
+        $query = DB::table('printings');
+        $query->where('printings.latest_status', 'print_success');
+
+        // somente as impressões do usuário em questão
+        if ($user) {
+            $query->where('printings.user_id', $user->id);
+        }
+
+        if ($printer) {
+            // contabiliza todas as impressões em todas as impressoras do respectivo usuário
+            $query->join('printers', 'printings.printer_id', '=', 'printers.id');
+
+            // considerando impressões das impressoras pertencentes a mesma regra
+            if ($printer->rule) {
+                $query->where('printers.rule_id', $printer->rule->id);
+            }
+        }
+
+        // somente impressões do mês ou do dia
+        if ($period == 'Mensal') {
+            $query->whereYear('printings.created_at', '=', date('Y'));
+            $query->whereMonth('printings.created_at', '=', date('n'));
+        } elseif ($period == 'Diário') {
+            $query->whereDate('printings.created_at', Carbon::today());
+        }
+
+        return $query->sum(DB::raw('printings.pages*printings.copies'));
     }
 }
