+++ resources/views/printings/partials/printings_quantities.blade.php
@@ -1,5 +1,5 @@
 <div align="center">
   Total de impressões:</b> {{ $quantities['Total'] ?? '' }} <br>
-  Impressões de hoje: {{ $quantities['Diário'] ?? '' }} <br>
+  Impressões de hoje: {{ $quantities['Diario'] ?? '' }} <br>
   <p style="color:red;">Impressões neste mês:</b> {{ $quantities['Mensal'] ?? '' }}</p>
 </div>
